Removed the need for Microsoft.VisualBasic.Compability but creating a code stub for VB6. The only two functions that are actually used are CopyArray and Format. The other functions that were causing compilation errors I've removed (they weren't called, and they didn't work!)
I've also had to upgrade the version of the ADODB in both the server DLL and the web site. This was because, well not putting too finer point, just didn't run under 64 bit. I've read around Google a bit and everything seems to point that it should be perfectly compatable as its still a com component. Moving to the ADO.NET library would break all of our database connecttion code.

Had to modify finds.ascx too because of the subtle difference with the new version of ADO and how it deals with dbnulls (I think anyway, it could just have been an error from task-32 where regional date settings has been put in.
The copyArray has been swappped out to a clone function which leaves the arrays exactly the same, any changes here need to be done with care as .net and vb6 start arrays at 1 and 0 accordingly (or is it the other way round).

Also of note is that we were using mm/dd/yyyy as a date format string. This really should be  MM/dd/yyyy as little mm gives us the month.