JIRA-2787 - Changes following code review. First issue was the bloody frmUseful again after you've gone into a definition and then runm a report. Put a reference to get frmuseful from the calendarreportframe at top of the js file.
Secondly there were some unhandled nulls that I've now wrapped up in isdbnull checks.

Also sorted out the closeclick problem if no records are returned after running a report.

As a cautionary note the "Absence By Department" report you mentioned doesn't run in 5.1 or in Data Manager - this is because the definition is corrupted (the sort order has a reference to course records for some reason)

jQuery1.51 removed from project as part of this check-in for some reason the IDE has got clever. (or has it?)