There was a problem publishing to a Windows 2008 box - namely it didn't like v6.1 of the ActievX Data Objects. It was manifesting itself as an error cannot convert connection to ado.connectionclass when you attempted to log it. Other versions caused an error cannot convert internalparameterclass when using the system generally Through a process of trial and error (and assistance by our firneds at Google) I selected 2.7 as this seems to be a balnace between the modern operating systems and some of the legacy ones we have to support. I do however have a feeling that this may get revisited at some point.
I tried during this to also remove the liberal use of CreateObject("ado blah blah") that we use throughout the application as this should improve performance, however only certain areas of the app regonise the adodb class library (vb code behind some ascx inline does not). I think this is because we have some dodgy dependency code in here somewhere. Ideally we should come back to this, but for the moment it will have to remain as is.

Also changed the compilation options to compile the server DLL to AnyCPU rather than x86 as this too was throwing warnings when compiling.