{18/02/2010 17:04:08~18/02/2010 17:04:10}JIRA-786 - Performance loading designer.
Not lightening, but should be better.

A combination of not using the ado default properties (.value), doing all of the setControlOrder in one big hit instead of each control seperately, and issue a move command instead of 4 moves for top, left, height and width.

Sounds like a lot of hashing round, but its basically just more efficient coding, but keeping the logic the same-ish.