JIRA-2861 - Partial fix. The problem was the udtab_ functions were addressing parent columns directly. This was fine and dandy unless that column was calculated and that that calculation had a dependency on the child table in question. If that was the case it was taking the value from before the save took place. Clear? No probably not. Let me try and give an example...
Parent table (personnel) has a current salary and a previous salary field. The current salary gets the top line of history and grabs the salary value (that worked). The previous salary would attempt to get the second or third line of salary. However when inserting a new salary record the trigger would be looking at the wrong record hence grabbing the value before the insert (an old value).

The fix is to embed the previous salary calc into the udftab_ function. Several caviats here though.

1) Sometimes we reference the udfcalc_ sometimes the expression will get converted to straight TSQL depending on whether that calc is dependent on the IsComplex flag (complex in this instance being defined as whether the calc references columns in other tables)

2) In order to avoid recursion a trigger on a table only fires once. So the salary trigger will fire first updating all its calcs. Then the personnel will fire. The personnel trigger knows not to refire the salary trigger. This is where we will have some issues. In JIRA example given (customer: Croda) in addition to the previous and current columns described above they also have a last increase amount on the salary record. In order for this to work the salary trigger would fire first, then personnel. Since the salary won't fire again the last increase amount on the salary record won't get updated. 
I've had a chat with Mark E and the friendly people over in support and we've agreed that this is a rubbish way of customising a system because it would not return the expected results even if this part of the "fault" was fixed as its getting the previous value from personnel records which is only ever the latest previous value and so using this in historic salary records is going to produce erroneous results. (I hope I've explained that - its complex I know)

3) The JIRA mentions a problem with updating dependents history when an absence is entered. This will require a separate code fix and so is not in this changeset.

Thankyou for listening.