JIRA-2823
First part of this fix is to distinguish between the app_path and a log file path. We now write the log files to the users temp folder instead of the app_path as we can't rely on having write access to the app_path. Run this past support and IT and this is acceptable.

Second part deals with network connectivity. We now have a timer hitting the database every 10 seconds which should keep the connection to the server alive. This should address the original issue as customers were only reporting the problem if the machine had been left idle (overnight usually) If an network error occurs during this poll job it will prompt the user to reconnect. If that fails then it alerts the user, says they've lost their changes and then graciously boots them out. The timer is disabled when you start saving the system manager as reconnection is not possible at this point because we're in the middle of a transaction. Trying to reconnect at this point would be very bad news.

This is not bulletproof by any means. There are still ways you could receive misc object variables not found error messages as the poll job only goes every 10 seconds, so if you are loading columns, editing screens, or generally navigating around the System when your connection dies its in the lap of the gods as to whether this time event or some other peice of code will connect to the database next. Putting a check in before every call to the database would be simple to do if we had wrapped up all our database calls through one function. We haven't, so it isn't!


Instructions for test.

Leave system idle for a long time (either sit staring at your screen for the whole day or leave overnight)
Yank the network cable out - plug back into to test reconnection then generally use the system manager to check everything is OK.
Yank the network cable out - do not plug back in. Are you kicked out graciously
Run system manager as a non administrator. Are log files created in the user temp folder.