JIRA-2821 - The override filter was being applied to all tables in the record profile including the child tables. This was the reason why the prompt would appear multiple times. Because of this all filters on child table were being ignored so its highly likely that erroneous results were being prodcued even if the profile appeared to run correctly.
The record profile can be run in three "modes". Running from the reports menu, from an individual record or as a report pack.
In "normal" mode all the specifed filters are included.
In "individual" mode only the current record is included but all filters on child tables are applied.
In "report pack" mode if there is an override filter use that otherwise use the specified filter from the definition. Again all child filters are applied irrespective.