JIRA-1789 Not sure why this is so difficult to reproduce - looking at the code I'd say that if you don't have the /printer property on the command line it'll always hang. I can now make it happen at will on my machine (Windows 7).
First a little back history. When we print something from VB6 we can only ever sent output to the default printer (back in those days you'd have to have performed some kind of sexual favours to the IT department to get multiple printers on your PC). Therefore what we've hacked in place is to temporarily set the default printer to what the user would like and then at the end of the print job reset it. Now all of this code uses low level API which behave differently depending on operating system (may explain why its intermittent).

The reason for this part hanging is that there's a loop that keep trying to set the default printer back and it keeps looping to infinity if it can't. By not having the /printer property the startup code to identify the current default printer is not run and so it will continually try and set the default printer to vbNullString. That's the trouble with computers they're thick! Thick, but persistant!

Anyway, simple fix - just remove the loop.