JIRA-3111 - Error "Case expressions may only be nested to level 10."
It's possible to get a "Case expressions may only be nested to level 10." error message when saving the system manager if you have a particular combination of functions, calculations and calculated columns. This was noticed by Mark E when customising the standard system and swapping out calculations for calculated columns. It has also been an error noted occasionally on some customer systems and we've generally re-customised to workaround those situations as a fix has not been obvious. 

There are several permutations to make this happen, I've attached two examples in the screenshots. The first is a cut down version of the new standard system calculations which brought this error to our attention. Swapping the "is_current_position" column so it references the calculation instead of the column directly makes the error disappear, however due to HRPRO-3078 the running of this calc may return unexpected results. 

The second screenshot is one where I've tried to pair the error down to the bare bones and have the whole thing in one calculation rather than referencing other columns. This needs to be setup precisely like this (yes, it did take many hours to work this out).

SQL can only handle ten nested levels of case statements. When we generate the triggers and udfs we calculate how many nest levels we are at any one point and the minute it trips past 10 we break the expression into multiple components. This then marks this calculation as "Iscomplex" and forces it to be a udf instead of inline T-SQL in the trigger. We increment the case count on the following conditions: 
An If...Then...Else function 
An Is Field Empty/Populated 
An Extract characters from the Left/Right 
Is Between 
Maximum/Minimum 

Now added to that list is Parentheses. The reason for this is the condition part of If...Then...Else always need to evalute to a logic block which parentheses cannot be guaranteed to return hence we have to wrap up another case level. In that logic block we have to do quite a bit of processing as the UI for the expression builder allows quite a flexible amount expressions, for example a user could have 
IsLogicField 
IsLogicField = LogicalExpression 
LogicalExpression 
NOT logicalExpression OR IsLogicField 
NOT logicalExpression = LogicField OR IsLogicField 
CharacterField = CharacterField 
NumericFunction > 0 

Lots options here. The comparision operator is not always there, but it must still evaluate to logic condition and so the system framework is wrapping in an additional CASE statement that was not getting added to the CaseCount and so could potentially blow up if it encountered the right set of conditions and parentheses. 

In this changeset I've also removed the DepthCharge variable as it was doing nothing other than making the code a bit more unreadable!

Update script amended to add the casecount to the parentheses function. This will need running against the db before this fix will work.

Added some semi-colons to the end of generated statements
1) so the beautifier can run against the code and
2) Future proofing SQL.