Changes to login process. Previously it was making numerous round trips to build the column and table privelidges collections. Now altered round to grab the data once and process on the client side.
Part of this work involves changing the cTablePrividleges to a standard .NET collection with the associated extension file. Reasons for this is performance and readability. Next task is to clear out the non-stadndard way in which we call the collections.

There are changes to the stored procs which will need running against any databases that uses this version of the code. A later deployment task will ensure this is done as part of the build but for the moment its a manual task.

Swapoed the connection property to be a adodb.connection class instead of object. Using objects should be used with caution as they are memory hogs. One day it would be nice to get rid of them from the server dll for good.