JIRA - 3055 - Added cost center to contract message. Needs scripts running on both nfp and std databases for fix to work.
Note #1 : The message from std to nfp is going to get ignored as most of the fields in nfp are calculated. 

Note #2: The conversion of the Appointments.Duty_type  to Salary.Grade column could produce binary data being truncated as one is 40 charcaters long and the other 6 in our standard systems.