JIRA-2998 - Formatting of homeAddress node in the staffchange.
The connector is now ready but we will not be able to delete and address because the message will get blocked by the general fusion schema validator (Validate in Fusion.Core.MessageValidators.SchemaValidators.cs) because it has been set to ignoreWhiteSpace. I could override this in our connector however I feel its key that each of the connectors uses the same schema validation code or else we are liable to get corrupt messages hurtling around. This will probably get found during integration testing and may possibly invlove changing the core code.

Another problem that seems to reside in the core code is to do with the stateRepublisher adding carriage returns when it republishes addressline2, 3, 4 and 5 which is causing the original reported JIRA issue.