Two fixes needed. Firstly (util_run_calendar_report_main) the session connection needs to be passed using CallByName instead of directly accessing the property. This sometimes worked and sometimes didn't depending on if you were running through the debugger or not. Made it safe by using CallByName.
When promptedValues was orginally upgraded the div parameter on getForm was not implemented and so it found the frmPrompted wherever it was on the page. However this is njo longer the case and so I've directly referenced it instead. In an ideal world I would have ensured that I just change the div parameter however frmPrompted is one of those "beautiful" forms that sometimes appears in the optionframe and sometimes in the workframe depending on whether you are running standard reports/coming from the navigation frame or running from a recedit screen. (There's a property in the form which says which one its running in - but obviously this is of no bloody use if you don't know ehere to look in the first place!!!!)

Also this is about my third attempt to check this stuff in because I've had to rollback a previous changeset because of a stupid typo. (entering a password on one monitor accedentally getting put into the function name in another monitor). TFS is totally anal about rolling back files.