This is a bit of a halfway house fix until we complete the whole deployment task.
As it stands at the moment the stored procs in the database project are all set to 'create' and there's some black magic code running in the background that decides whether a object needs publishing or not. So what I've had to do is to publish to a blank database and then use management studio to script with the 'DROP AND CREATE' of all objects with the prefix spASRInt and sp_ASRInt. 

Also at this stage there is no version checking or setting of any kind, that is because we are still trying to use 5.1 and 8.0 databases side by side. At some point we are going to have to split this, but this too can be reviewed when we look at automated build/deployment. (Its made evgen more confusing by QA testing an interim 5.2 build at the same time!) As a result of no chcking you may get some warnings about dropping no existant procs if you pick a database that was not already running v8.0 (they can be ignored as the new functions will get created anyway)

To test run the updateScript8.0.9 on a database and review the associated work items. Very messy I know, but should work for this sprint.