JIRA-3157. The problem was having a calculated column that gets a value from a child table but which also has a filter on it that looks that a column thats from another parent.
Background. (5.2 original fix)
When building a udf we look at any column and if its a calculated column instead of just referencing that column (because it won't have the latest value in the db at the point the trigger fires) we swap it out for the associated calculation. This is fine and dandy expect if that column goes off an references columns in other tables that aren't necessarily in the base table.

So we do some jiggery pokery with swapping "base." with "tablename." so that we can use the inlinecode rather than the callingcode. (the rest of sqladdcolumn and builddependencies adds the table relations (hopefully!))

This changeset fix
All that this particular fix does is ensures that any IDs are still linked to the base table because this is the only table that will hold a reference to all the parents. 

As for the JIRA issue that related to Merlin I make no guarantee that the results returns from this calculation will make sense because the logic for this calc is dubious at best.

(Test harness also updated to reflect different default database)