There's a lot of changes in here so make yourself comfy.
Took out the ClientDLL and replaced with js one (output features not yet implemented this is merely a code stub

Some variables had no declared type and object was assumed, this is very resource hungry as its processed at runtime so swapped for compile-time object types. (This is good practice and when I get five minutes (which I won't!) might be worth rolling into all other areas of the system where performance is under scrutiny

Make several FORM and INPUT declarations to be lowercase as is now the standard

Killed the ssHiddenGrid, was only used for printing which is superseded by the output options anyway.

Removed a couple of redundant functions and variables from CustomReportsRun.vb

Swapped some functions to be subs in CustomReportsRun.vb

Removed the activeX handlers (obviously. I only mention it because I left them in originally and then spent the next hour and a half working out why the bloody browser kept crashing. Idiot!)

Building the strings for the grid/table have been altered to change from vbTab as the delimter to be <td></td>

The arrays need to be zero bound rather than starting at 1, this goes with swapping some of the datatypes from object types with redim preserves to using an arraylist. Should give a performance boost on larger reports

Removed references to mlngColWidth as this is no longer applicable sending out to a table

To test try all custom reports with different datatypes and referencing child/parent tables. Also had to make significant logic changes to the way the group with next works so this needs reviewing too.

Along the way of testing this myself I've fixed up the following JIRA faults.

2796 - Needed a tostring as a default property in formatdata and format summary functions
2974 - Accidentally fixed along with changes in format data
2973 - Caption was missing. Added just after reportdata_window_onload is called.

Work still todo.
Styling this report - at present it looks different to a cross tab and different again to defsel. I've put the classes in and a basic stylesheet but someone needs to attack this with some crayons.
Hook in the output options.