JIRA-3076 - Output and cancel buttons missing from calendar reports.
Yup - the util_run_calendarreport_output had not been converted. As this was such a small pitiful little view I've incorporated the code directly into the main form as a table. 

As part of this task I've also tidied up the output form for all the other report types as they were not pulling the default values through.

THe submitForm has been tweaked so that we can point an ajax query as a specific div (yes there was code in there before - but it was "functionally compromised"