Reported part of the bug was caused by not initialising the output email name in the stored proc. This will need running against any database where this fix is tested (update script to be completed later in this sprint)
While I was in here I noticed that the subsequent disabling of controls was not happening. This was caused by having an extra tab in the usernmame and access variables when the defsel grid was built (an overhang from the activeX removal in defsel)

Also reszied some of the grids so they look a little better on screen.

Other misc fixes in this are removal of superfluos mouseover and mouseout events and renaming a couple of script function as they might be clashing with other report types.

Also swapped out the createobject to be a nice stronly typed control, hopefully giving us a bit of a performance boost (at the very least intellisense now works!)