Removed all references to gbEnableUDFFunctions as this is now redundant. This was a legacy variable when we had to support SQL2000. This is now always set to true.
Swapped the sourcetables array from object array to integer array. We need to do this because .NET has a habit of detecting what an object type is an automatically casting it to that type which can cause problems if we're not specific.

The udffilter code was not being generated properly because of the difference between .NET and classic VB in the way that arrays are zero bounded. Probably would have been a good idea to swap this to an arraylist as the reszing would have been automatic but that would have taken a bit too long and been a higher risk. I've swapped the code to use length instead of UBound as this is generally more in line with .net commands and according to Google is a little more performant, however its important to note the differences ubound returns the upper limit of one of the dimensions, whereas length returns the totla amount of elements which could be differnt in a multi-imensional array. Its safe for this instance, but should not be considered as a general rule for all conversions.
http://support.microsoft.com/kb/311333

Removed some ancient commented out code for purposes of readability

Removed some unused variables

Moved UDFFunctions to be a shared function as there were changes necessary to support zero bounded arrays and I didn't want to have to replicate that work five or six times. Rewrote to use a try-catch instead of a badly performing "on error". This is an opportunity to imporve preformance through the server dll as on error starts to degrade if the scope is over twenty or so lines (below that its much the same)

We deinately need to sort out not having a progress bar for running these reports as it looking kinda horrible when running big or reasonably complex reports.

JIRA references 
2991 - udf broken
3180 - Output differences between intranet and DM