Changed some of the objuect type refernces to use adodb.parameter, command and recordset for performance and readability and intellisense.
The udffunctions needed a bit of error handling.

The data-framesource was not being set for picklistmain hence some of the objects were still in context but had been cleared from the dom hence the reported error.

Got rid of util_dialog_picklist as it was just an uneccessary step. Rejigged some of the form code to handle this.

We still have to hide the workframeset when popping up the picklist record selection because of the pesky activeX bug always displaying on the top. This ideally could do with fixing up at some point.

Cleaned up a few bit of old commented out code

Made some html tags lower case

Tasks still to do
We really could do with moving a load of the javascript code from the aspx pages to a seperate .js file. I haven't done this because we are only one day out of the sprint build.

All of the parameters on the adodb.parameter objects need to be strongly typed (it all works but makes resharper unhappy)