JIRA-3153 - If omitted nodes then take the existing value.
I have not catered for multi master messages that are hoofed through the republisher (staffChange, staffSkillChange, contactChange) as they will get all 
of the nodes added as part of the state republisher. (or should have at least)

The only ones where this mechanism has been implemented is (staffLegalDocumentChange, staffContractChange)

StaffPictureChange and TimsheetSubmission have not been implemented as all of the fields have a minoccurs of 1

Also started doing a few checks for saga.