Part of this fault will also start tidying up issues with connections timeouts and tripping over each other. This is because there were some shared subs which meant that the connection was shared per server session rather than client one (whoops). (Server DLL changes)
Then removed some of the createobjects to use adodb.command directly as this means that variablkes can be bound at compile time. Can't say for sure exactly how much this will improve performance but according to every rule in the book it does, it should also reduce the memory footprint/leak although this is probably only noticeable when it gets put under load. (Note: there are lots of other places this change needs rolling into, but will form part of a later changeset so as to not overcomplicate this one)