Swapped the building of the histroy screens and child screens to use a custom class and then iterate through avoiding the use of a keyset cursors as this removed the millions of sp_cursorfetch everytime you log in.
Cleared the menu on a menu_refresh to avoid doubling up of menu items./

Need to generally go through the system at a later date and put in some proper error handling as this is pretty rubbish at the moment.

Looks like this check-in has removed a large chunk of code at the bottom. The merge of the check-in should resolve. Hopefully.