Few things with this check-in. The simple stuff first - removed some old useless commented out code.
Secondly there were two stored procs being called before sp_ASRIntGetFindRecords3 which were surprsingly heavy considering all they were retrieving was flags of which columns were set to blank if zero and which has a thousand seperator as they both went off and scanned the sysprotects table. I've now got rid of the calls to them entirely and altered the main find procedure to return 2 recordsets, one for some of the parameters and one for the data. 

Why didn't I just return the values in the output paramters of the command object I hear you ask - well good ol' ADO does not set the parameters values until the associated recordset is complete (a common problem which has bugged the Intranet for a while). So I'm returning the two bits of new information from the stored proc as a first resultset. (I haven't altered the other parameters because I want to keep changes to a minimum at the moment, but for readability and consistency its probably a good idea to do it at some point (either that or scrap the whole stored proc and put a control in that bind directly to the database!!!))