Moved the formatting logic to the read recordset level. Swapped main code to a string format and removbed the duplication for formatting numerics.
Note 1: If the length of a numeric column is too long it now displays #### instead of just trimming the data. It might be different to before and the DM but this is surely a better design.
Note 2: Numeric columns still not formatted on the right correctly). Needs a bit of tidying up on the stylesheet.
Note 3: There's still more preformance improvements that can be done in that formatgrid function