Bumped up the version to 8. (Does not include the automated database upgrade script - that will be a seperate task - run scripts in the databse folder if running on a genuine 5.2 database)
In my opinion the icon looks a little furry, but this is the only standard one we have at the moment, possily get our web designeer to have a look at)

Have run the script against npg_openhr8 and npg_openhr8_2 so to test you'll need another db.

Fat clients modified because there was a stupid bug that kept looping around if you skipped a version (i.e. 6 and 7)