Ribbon not correctly displaying reports when in record edit.
Note: This code is purely about the buttons displaying - not the actual running of these. (Will need addressing immediately after this alpha demo (maybe before?)

Clicking on close does not bring back the recordedit screen if you selected mail merge or calendar reports. The reason for this is that defsel is getting loaded into the workframe rather than the optionsframe and hecne the record edit session is being "zapped". Needs to be altered to load defsel in optionsframe.

The bradford factor run is disbaled because at the moment it blows up. This also needs addressing at some point.

Running the mail merge is permissions problem. Again I think this is in part to the defsel being in the workframe rather than the optionsfarme