JIRA-3239 - Now checks for duplicate messages before sending them. Also checks for processed message when recieving data.
Implemented a baseclass for data objects and moved some of the code from the builder to the data class so as to avoid code duplication.

When a message is recieved from the bus it regenerates the message to pretend as if we had sent it.

God knows what consequences this could have...