JIRA-3158 - Process the staffTimesheetSubmission message.
What a right pain in the backside this one was! Firstly I've had to amend the corporate test harness to process this message. So if you get latest of the core code from SVN you may well lose the ability to test this message.

Secondly I've added a default service name to the project - this is so that when running in code the conector is listening on the correct queue. In an installation this service name would be specified when installing nservicebus.host.exe as a service.

Then there were some wrong parameters on the xml definitions.

Lastly the sql scripts were trying to insert into the ID field.

I should note that up until today this message has been impossibe to test as we've not had a harness, so its not surprising there are a few quirks in here.