Styling added to the popups and logic added to prevent hitting enter when you shouldn't.
I noticed as part of doing this that adding another reference to the jquery bundles causes problems if that form goes out of scope, almost as if its unloading things.

Fixed up numerous other mini-bugs which I noticed in part of this mostly to do with not having proper object types defined. (I'll track through JIRA and see if I can knock some of those off too)