Had a problem with moving around the system generally as we are defintion var frmUseful, frmDefinition etc at the top of all the .js files. The problem occurs when they go out of scope, which is what was happening when swiucthing between calcs.filters/mail merge and picklists.
So to "fix" this I've actally made things a bit worse (sorry) by adding the picklisty stuff into its own js file and declarion those variables at the top. Ideally we should not be having "global" variables at the js files but that would take quite a bit of code changes. I'll leave that to someone else...

Remember I've added something to the bundle (yuck) so remember rebuild the solution and ctrl-f5...