Moved some extra metadata to be loaded at runtime (maybe more to come operators, operator parameters and function paramteres).
Also swapped out the getRecords to be openrecordset with a firehose cursor to try and increase performance. This does now mean we are dependant on having mars on the connection string (I think we were anyway)

Cleaned up some string handling. Lots of pointless & removed and moved some stuff to string.format where it was simple to do so.

Also fixed up a small bugette when there was a 0 passed getcolumnname.

As this change affects all areas you may want to "consider" all report types and utilities.