More performance gains, this time with an eye to boost performance of the dashboard (in particular database values). Swapped more code to use firehoses and added operators and operator types to the cached list of objects. (We may want to cosnider the populateData function as a target for optimisation with NHibernate or stream it as one round trip to the server?)
Split the ExpresswionValueTypes into its own class so that it cane be exposed in the other object classes. This has resulted in numerous small changes in files to remove the Declarations.

Added some isnull checks in the  get charting data (not sure that is still necessary as I was chasing another big at the time, but it seems a good precaution to leave it there)

Added the database values in LinksMain - lets see what the performance is like on systems other than our own?

Right, I think that's it, I'm off to catch last orders...