Noted when testing some of the very large and complex reports setup by QA (qasrvr01 - big calendar report run with caution).
Was devilishly slow. This was partly due to the huge amount of bank holidays (20+ years worth) and the innefficient string builing). Move the code out to a string builder and its now almost useable for this level of data (minute to load as opposed to timing out after 20)

This is still not ideal and I have some ideas on how to improve (remove class off the input elements (alter site.master))
Switch the data to be a json output string instead of input tags (bigger job)
Rewrite the whole bloody thing!

P.S. also removed a _ from one of the string being built which unsurprisingly blew up.