Removed all the ActiveX for Calendar report running. There have been extensive changes for this.
1) Controls removed and replavced with a DayPilot schedule control. This expects the data in the form of a DataTable and so all of the existing code to write out input controls which are then read by the client into the ActiveX have been removed. A conversion of ado recordset to ado.net datatable has been written. Ideally in the newar future the whole calendar report data compilation will be moved over to ado.net hence removing this step (and increasing performance)

2) The options grid is no longer displayed. This was done simply for convenience and the belief that these options are not essential for the moment (as agreed with Mark E and Harry C). The SQL code is still in place for when/if these options are deemed to be required once again.

3) The breakdown form has been completely trashed and rewritten using proper (ish) MVC. The DayPilot control only passes back the event ID and not the individual date when an event is clicked and so we can't pick up on any change in the working pattern during the event period. Discuss whether this is a problem. 

4) Several forms have been removed - critical forms and processing has been combined into other forms.

This code is checked in with a styling task still outstanding (task 2064). Bits that need doing for this. Add themeroller. Tidy up resizing? Made it look pretty. Format the event breakdown.

I've used the off the shelf daypilotLite control, but if we want to do anything fancy we may have to alter it. We are allowed, but that could mean upgrade issues down the line. They do have a pro version with lots of other fancy bits in it which we could look at using for future releases. (Things like edit events, gantt charts, day/week/year views etc. I've checked this code into SourceSafe (but I wouldn't bother reviewing it as it's 3rd party)

Its still a bit of a mess as its an upgrade of old classic ASP. With a time of time this can be cleaner.