Have moved the running of the mail merge from the client to the server. The steps to do this was to take the ActiveX vb6 code use the project upgrade wizard in VS2008 (they removed it in 2010 onwards) and then remove the errors. This file is included as MailMergeClient.vb. This will probabbly get tidied up when I get round to resolving task 2701.
There was a lot of code for messageboxes in there which have been replaced with strings aand returned to the calling function, but there's still some tidying up to do. Not least it leaving the reportframe popped up after running a mail merge.

When you select display on screen in the defintion the functionality is to now to download a genberated file instead of loading Word directly. With Chrome you can get the browser to default an automated action to open for certain file types, so with Chrome you can setup your PC to emulate pretty much exact functionality of 5.2

There are some bits of functionality that are no longer in the mail merge run
1) Prompting the user before running the merge to ask if they want to alter the template document. Tricky to implement because it would mean streaming a document down to the client and then getting them to upload again.

2) The file locations would previously have been setup from the clients perspctive (i.e. c:, d: etc). There is no guarantee (in fact its unlikely) that the server will have the same configuarion, so upgraded databases will need some thinking about (OPENHR8 has an x: and y: drive defined the same as our client pcs, or should be anyway) so we can run, but some "consideration" will be required before we hoof over to QA.

3) Locale settings not implemented.

4) Automatic Emailing out of the merged document not implemented.

There is also a rqeuirement for Office to now be installed on the IIS box. I can see this being a bit of issue for some customers/IT departments as they like to keep the IIS box as light and secure as possible, particularily if its accessible to the outside world. This would involve a bit of rewriting of that mailmeregclient dll to rely on a control other than the Word app. Infragistics supply such a control (as do a whole host of 3rd party tools for that matter)

All of the above points have been raised as separate TFS task items, maybe to look at in this sprint? Or maybe not?